APIs/akeneo.com/1.0.0/python-aiohttp-client/tests/test_product_media_file_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.get_media_files_code200_response import GetMediaFilesCode200Response
from openapi_server.models.media_files import MediaFiles
from openapi_server.models.post_media_files_request import PostMediaFilesRequest
from openapi_server.models.post_token400_response import PostToken400Response


pytestmark = pytest.mark.asyncio

async def test_get_media_files(client):
    """Test case for get_media_files

    Get a list of product media files
    """
    params = [('page', 1),
                    ('limit', 10),
                    ('with_count', False)]
    headers = { 
        'Accept': 'application/json',
    }
    response = await client.request(
        method='GET',
        path='/api/rest/v1/media-files',
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_get_media_files_code(client):
    """Test case for get_media_files_code

    Get a product media file
    """
    headers = { 
        'Accept': 'application/json',
    }
    response = await client.request(
        method='GET',
        path='/api/rest/v1/media-files/{code}'.format(code='code_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_get_media_files_code_download(client):
    """Test case for get_media_files_code_download

    Download a product media file
    """
    headers = { 
        'Accept': 'application/json',
    }
    response = await client.request(
        method='GET',
        path='/api/rest/v1/media-files/{code}/download'.format(code='code_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

@pytest.mark.skip("*/* not supported by Connexion. Use application/json instead. See https://github.com/zalando/connexion/pull/760")
async def test_post_media_files(client):
    """Test case for post_media_files

    Create a new product media file
    """
    body = openapi_server.PostMediaFilesRequest()
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'content_type': 'content_type_example',
    }
    response = await client.request(
        method='POST',
        path='/api/rest/v1/media-files',
        headers=headers,
        json=body,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

